@@ -5,8 +5,8 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"golang.org/x/exp/slices"
 	"net/url"
+	"slices"
 	"strconv"
 	"strings"
 	"time"
