@@ -5,6 +5,8 @@ package connectors
 import (
 	"HalogenGhostCore/core"
 	"encoding/base64"
+	"fmt"
+	"log"
 	"math"
 	"math/rand"
 	"net/url"
@@ -13,7 +15,53 @@ import (
 	"time"
 )
 
-var loc, _ = time.LoadLocation("Europe/Moscow")
+type GDConnector struct {
+	output string
+}
+
+func (c *GDConnector) Output() string {
+	log.Println("Output: " + c.output)
+	return c.output
+}
+
+func (c *GDConnector) Error(code string, reason string) {
+	log.Println("Error: " + code + " " + reason)
+	c.output = code
+}
+
+func (c *GDConnector) Success(message string) {
+	c.output = "1"
+}
+
+func (c *GDConnector) Account_Sync(savedata string) {
+	c.output = savedata + ";21;30;a;a"
+}
+
+func (c *GDConnector) Account_Login(uid int) {
+	c.output = fmt.Sprintf("%d,%d", uid, uid)
+}
+
+func (c *GDConnector) Comment_AccountGet(comments []core.CComment, count int, page int) {
+	if len(comments) == 0 {
+		c.output = "#0:0:0"
+	} else {
+		for _, comm := range comments {
+			c.output += c.getAccountComment(comm)
+		}
+		c.output = fmt.Sprintf("%s#%d:%d:10", c.output[:len(c.output)-1], count, page*10)
+	}
+}
+
+// GetAccountComment used to retrieve account comments (iterative, w/o hash)
+func (c *GDConnector) getAccountComment(comment core.CComment) string {
+	s := strconv.Itoa
+	t, err := time.ParseInLocation("2006-01-02 15:04:05", comment.PostedTime, loc)
+	if err != nil {
+		t = time.Now()
+	}
+	age := core.GetDateAgo(t.Unix())
+	return "2~" + comment.Comment + "~3~" + s(comment.Uid) + "~4~" + s(comment.Likes) + "~5~0~6~" + s(comment.Id) + "~7~" + s(core.ToInt(comment.IsSpam)) + "~9~" + age + "|"
+}
 
 // GetUserProfile used at getUserInfo (w/o trailing hash)
 func GetUserProfile(acc core.CAccount, isFriend bool) string {
@@ -47,17 +95,6 @@ func UserSearchItem(acc core.CAccount) string {
 	return "1:" + acc.Uname + ":2:" + s(acc.Uid) + ":3:" + s(acc.Stars) + ":4:" + s(acc.Demons) + ":8:" + s(acc.CPoints) + ":9:" + s(acc.GetShownIcon()) +
 		":10:" + s(acc.ColorPrimary) + ":11:" + s(acc.ColorSecondary) + ":13:" + s(acc.Coins) + ":14:" + s(acc.IconType) + ":15:" + s(acc.Special) +
 		":16:" + s(acc.Uid) + ":17:" + s(acc.UCoins) + ":52:" + s(acc.Moons) + "#1:0:10"
-}
-
-// GetAccountComment used to retrieve account comments (iterative, w/o hash)
-func GetAccountComment(comment core.CComment) string {
-	s := strconv.Itoa
-	t, err := time.ParseInLocation("2006-01-02 15:04:05", comment.PostedTime, loc)
-	if err != nil {
-		t = time.Now()
-	}
-	age := core.GetDateAgo(t.Unix())
-	return "2~" + comment.Comment + "~3~" + s(comment.Uid) + "~4~" + s(comment.Likes) + "~5~0~6~" + s(comment.Id) + "~7~" + s(core.ToInt(comment.IsSpam)) + "~9~" + age + "|"
 }
 
 // GetLevelComment used to retrieve level comment (iterative, w/o hash)
