@@ -16,13 +16,13 @@ import (
 	"github.com/aws/aws-sdk-go/service/s3/s3manager"
 	"github.com/getsentry/sentry-go"
 	"github.com/go-redis/redis/v8"
-	"golang.org/x/exp/slices"
 	"html"
 	"io"
 	"math/rand"
 	"net/http"
 	"net/url"
 	"os"
+	"slices"
 	"strconv"
 	"strings"
 	"time"
